okay, this will touch most directories of Miro I just write one message for all of it I compiled everything after a make disclean - so I'm quite positive, that  everything still works anyhow, brace for impact
Documenting Log.h making it hopfully more usable for other developers

Trying to properly separate miroSvc.a from miro.so
-> It should be possible to link to miro.so without linking miroSvc.a
Linking miro.so with miroIDL.so as it is needed anyways.
-> It should be possible to link to miro.so without linking miroIDL.so
Propagating this change through the sources

Splitting up Parameters.xml into Parameters.xml and SvcParameters.xml.
This way we can skip some classes from inclusion into miro.so
Additionally we may need to recompile less on changes in
SvcParameters.xml
Propagating this change through the souces

Adding and removing headers for compability with ACE 5.4/TAO 1.4
We need to make sure Miro compiles with the current stable release of
ACE/TAO. - There is a vired bug in ACE 5.3/TAO 1.4/g++ 3.3.1 witch
requires to include cmath before OS.h. So this workaround was
propagated through the sources, too.

added a chapter from Gerd about autoconf and automake to the manual
appendix - moved the part from the writing_a_client.tex about
Makefiles to this chapter, too.

silencing cvs by adding even more .cvsignore entries. It's fascitating
how mutch you can ignore while building miro...

removed PositionTracker.h form src/miro as it is actullay substituted
by OdometryTracking, which in turn needs to be unified with the
version residing in the Nix repository.

+++ trunk/src/video/VideoDeviceBTTV.cpp
@@ -8,6 +8,7 @@
 // $Id$
 //
 //////////////////////////////////////////////////////////////////////////////
+#include <cmath>
 #include <ace/OS.h>
 
 #include "VideoDeviceBTTV.h"
@@ -153,7 +154,7 @@ namespace Video
     std::cout << "buffersize/frame: " << gb_buffers.size/gb_buffers.frames << std::endl;
     std::cout << "frames: " << gb_buffers.frames << std::endl;
 
-    if (gb_buffers.frames < params_->buffers) {
+    if (gb_buffers.frames < (int)params_->buffers) {
       throw Miro::Exception("Number of requested buffers in the configuration\n" \
 			    "exceeds the number of frames available by the device.");
     }
